refactor: combine test files
