+++ tests/e2e/e2e_gatewayapi_test.go
@@ -36,6 +36,94 @@ const (
 	nsUDPSvc    = "udp"
 )
 
+var _ = FSMDescribe("Test traffic routing by FSM Gateway with trafficInterceptionMode(PodLevel)",
+	FSMDescribeInfo{
+		Tier:   1,
+		Bucket: 6,
+		OS:     OSCrossPlatform,
+	},
+	func() {
+		Context("Test traffic from client to backend service routing by FSM Gateway", func() {
+			It("allow traffic of multiple protocols through Gateway", func() {
+				// Install FSM
+				installOpts := Td.GetFSMInstallOpts()
+				installOpts.EnableIngress = false
+				installOpts.EnableGateway = true
+				installOpts.EnableServiceLB = true
+				installOpts.TrafficInterceptionMode = constants.TrafficInterceptionModePodLevel
+
+				Expect(Td.InstallFSM(installOpts)).To(Succeed())
+				Expect(Td.WaitForPodsRunningReady(Td.FsmNamespace, 3, &metav1.LabelSelector{
+					MatchLabels: map[string]string{
+						"app.kubernetes.io/instance": "fsm",
+						"app.kubernetes.io/name":     "flomesh.io",
+					},
+				})).To(Succeed())
+
+				testDeployFSMGateway()
+
+				testFSMGatewayHTTPTrafficSameNamespace()
+				testFSMGatewayHTTPTrafficCrossNamespace()
+				testFSMGatewayHTTPSTraffic()
+				testFSMGatewayTLSTerminate()
+				testFSMGatewayTLSPassthrough()
+
+				testFSMGatewayGRPCTrafficSameNamespace()
+				testFSMGatewayGRPCTrafficCrossNamespace()
+				testFSMGatewayGRPCSTraffic()
+
+				testFSMGatewayTCPTrafficSameNamespace()
+				testFSMGatewayTCPTrafficCrossNamespace()
+				testFSMGatewayUDPTrafficSameNamespace()
+				testFSMGatewayUDPTrafficCrossNamespace()
+			})
+		})
+	})
+
+var _ = FSMDescribe("Test traffic routing by FSM Gateway with trafficInterceptionMode(NodeLevel)",
+	FSMDescribeInfo{
+		Tier:   1,
+		Bucket: 7,
+		OS:     OSCrossPlatform,
+	},
+	func() {
+		Context("Test traffic from client to backend service routing by FSM Gateway", func() {
+			It("allow traffic of multiple protocols through Gateway", func() {
+				// Install FSM
+				installOpts := Td.GetFSMInstallOpts()
+				installOpts.EnableIngress = false
+				installOpts.EnableGateway = true
+				installOpts.EnableServiceLB = true
+				installOpts.TrafficInterceptionMode = constants.TrafficInterceptionModeNodeLevel
+
+				Expect(Td.InstallFSM(installOpts)).To(Succeed())
+				Expect(Td.WaitForPodsRunningReady(Td.FsmNamespace, 3, &metav1.LabelSelector{
+					MatchLabels: map[string]string{
+						"app.kubernetes.io/instance": "fsm",
+						"app.kubernetes.io/name":     "flomesh.io",
+					},
+				})).To(Succeed())
+
+				testDeployFSMGateway()
+
+				testFSMGatewayHTTPTrafficSameNamespace()
+				testFSMGatewayHTTPTrafficCrossNamespace()
+				testFSMGatewayHTTPSTraffic()
+				testFSMGatewayTLSTerminate()
+				testFSMGatewayTLSPassthrough()
+
+				testFSMGatewayGRPCTrafficSameNamespace()
+				testFSMGatewayGRPCTrafficCrossNamespace()
+				testFSMGatewayGRPCSTraffic()
+
+				testFSMGatewayTCPTrafficSameNamespace()
+				testFSMGatewayTCPTrafficCrossNamespace()
+				testFSMGatewayUDPTrafficSameNamespace()
+				testFSMGatewayUDPTrafficCrossNamespace()
+			})
+		})
+	})
+
 func testDeployFSMGateway() {
 	// Create namespaces
 	Expect(Td.CreateNs(nsGateway, nil)).To(Succeed())
